minor fixes, renamed custom_button.dart file
