+++ frontend/lib/providers.dart
@@ -2,8 +2,8 @@ import 'package:flutter_riverpod/flutter_riverpod.dart';
 
 final connectionsFutureProvider = FutureProvider<List<String>>(
   (ref) async {
-    return await Future.delayed(
-        const Duration(seconds: 3), () => ["Aalto University", "Finland"]);
+    return await Future.delayed(const Duration(seconds: 2),
+        () => ["Aalto University", "Finland", "OP Financial Group"]);
   },
 );
