+++ frontend/lib/widgets/credential_card.dart
@@ -14,85 +14,78 @@ class CredentialCard extends ConsumerWidget {
   Widget build(BuildContext context, WidgetRef ref) {
     final height = MediaQuery.of(context).size.height;
 
-    return InkWell(
-        // TODO: implement onTap to expand the credentialCard
-        //onTap: () => ,
-        child: Padding(
+    return Padding(
       padding: const EdgeInsets.all(8),
-      child: Column(
-        children: [
-          Ink(
-            decoration: BoxDecoration(
-              color: const Color.fromARGB(125, 222, 222, 222),
-              borderRadius: BorderRadius.circular(8),
-            ),
-            height: max(height * 0.12, 50),
-            child: Row(
-              crossAxisAlignment: CrossAxisAlignment.center,
-              mainAxisAlignment: MainAxisAlignment.center,
-              children: [
-                Padding(
-                  padding: const EdgeInsets.all(16),
-                  child: ClipRRect(
-                    borderRadius: BorderRadius.circular(15.0),
-                    child: AspectRatio(
-                      aspectRatio: 1,
-                      child: Container(
-                        decoration: const BoxDecoration(
-                          image: DecorationImage(
-                            fit: BoxFit.fitHeight,
-                            alignment: FractionalOffset.center,
-                            image: AssetImage('assets/logos/findywallet.png'),
-                          ),
-                        ),
+      child: Container(
+        decoration: BoxDecoration(
+          color: const Color.fromARGB(125, 222, 222, 222),
+          borderRadius: BorderRadius.circular(8),
+        ),
+        height: max(height * 0.12, 50),
+        child: Row(
+          crossAxisAlignment: CrossAxisAlignment.center,
+          mainAxisAlignment: MainAxisAlignment.center,
+          children: [
+            Padding(
+              padding: const EdgeInsets.all(16),
+              child: ClipRRect(
+                borderRadius: BorderRadius.circular(15.0),
+                child: AspectRatio(
+                  aspectRatio: 1,
+                  child: Container(
+                    decoration: const BoxDecoration(
+                      image: DecorationImage(
+                        fit: BoxFit.fitHeight,
+                        alignment: FractionalOffset.center,
+                        image: AssetImage('assets/logos/findywallet.png'),
                       ),
                     ),
                   ),
                 ),
-                Expanded(
-                  child: Column(
-                    mainAxisAlignment: MainAxisAlignment.center,
-                    crossAxisAlignment: CrossAxisAlignment.start,
-                    children: [
-                      Padding(
-                        padding: const EdgeInsets.only(
-                          bottom: 7,
-                          top: 23,
-                        ),
-                        child: Text(
-                          name,
-                          textAlign: TextAlign.left,
-                          style: const TextStyle(
-                              fontWeight: FontWeight.w900, fontSize: 16),
-                        ),
-                      ),
-                      const Expanded(
-                        child: Padding(
-                          padding: EdgeInsets.only(bottom: 5),
-                          child: SingleChildScrollView(
-                            child: Text(
-                              "description",
-                              textAlign: TextAlign.left,
-                              style: TextStyle(fontWeight: FontWeight.w500),
-                            ),
-                          ),
-                        ),
-                      ),
-                    ],
-                  ),
-                ),
-                const Padding(
-                  padding: EdgeInsets.all(8),
-                  child: Icon(
-                    Icons.expand_more,
-                    size: 40,
-                  ),
-                ),
-              ],
+              ),
             ),
-          ),
-        ],
+            Expanded(
+              child: Column(
+                mainAxisAlignment: MainAxisAlignment.center,
+                crossAxisAlignment: CrossAxisAlignment.start,
+                children: [
+                  Padding(
+                    padding: const EdgeInsets.only(
+                      bottom: 7,
+                      top: 23,
+                    ),
+                    child: Text(
+                      name,
+                      textAlign: TextAlign.left,
+                      style: const TextStyle(
+                          fontWeight: FontWeight.w900, fontSize: 16),
+                    ),
+                  ),
+                  const Expanded(
+                    child: Padding(
+                      padding: EdgeInsets.only(bottom: 5),
+                      child: SingleChildScrollView(
+                        child: Text(
+                          "description",
+                          textAlign: TextAlign.left,
+                          style: TextStyle(fontWeight: FontWeight.w500),
+                        ),
+                      ),
+                    ),
+                  ),
+                ],
+              ),
+            ),
+            const Padding(
+              padding: EdgeInsets.all(8),
+              child: Icon(
+                Icons.expand_more,
+                size: 40,
+              ),
+            ),
+          ],
+        ),
       ),
-    ));
+    );
   }
 }
