+++ frontend/lib/widgets/navigation_screen_outline.dart
@@ -30,12 +30,15 @@ class _NavigationScreenOutlineState extends State<NavigationScreenOutline> {
         backgroundColor: Colors.purple,
         actions: [
           TextButton(
-              onPressed: () {
-                context.go('/');
-              },
-              style: style,
-              child: const Column(
-                  children: [Icon(Icons.logout_rounded), Text('Logout')]))
+            onPressed: () {
+              context.go('/');
+            },
+            style: style,
+            child: const Tooltip(
+              message: "Log out",
+              child: Icon(Icons.logout_rounded),
+            ),
+          )
         ],
       ),
       body: SafeArea(
